+++ event-notificator/src/main/java/ru/haritonenko/eventnotificator/domain/EventNotification.java
@@ -3,9 +3,9 @@ package ru.haritonenko.eventnotificator.domain;
 import java.time.LocalDateTime;
 
 public record EventNotification(
-        Integer id,
-        Integer userId,
-        Integer eventId,
+        Long id,
+        Long userId,
+        Long eventId,
         LocalDateTime createdAt,
         boolean read,
         String message
